refactor: move searchTool and Search result to a MainPage folder
