+++ react-components/src/components/pages/MainPage/components/SearchTool.tsx
